requirements.txt
django-environ==0.4.0
django-static-sites==0.0.8
django-storages==1.4
markdown2==2.3.7
requests==2.25.1
